pyproject.toml
[project]
name = "procure-to-pay"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.8",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "gunicorn>=21.2.0",
    "pdfplumber>=0.11.8",
    "pillow>=12.0.0",
    "psycopg2-binary>=2.9.11",
    "pypdf2>=3.0.1",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.2.1",
    "weasyprint>=66.0",
]
